Adds a placeholder api call for demoing REST capabilities
